+++ src/revlanguage/functions/math/Func_diagonalMatrix.cpp
@@ -27,7 +27,7 @@ Func_diagonalMatrix* Func_diagonalMatrix::clone( void ) const
 RevPtr<RevVariable> Func_diagonalMatrix::execute( void )
 {
     
-    int n = static_cast<const Natural &>( args[0].getVariable()->getRevObject() ).getValue();
+    int n = (int)static_cast<const Natural &>( args[0].getVariable()->getRevObject() ).getValue();
     RevBayesCore::MatrixReal *m = new RevBayesCore::MatrixReal(n, n, 0.0);
     for (size_t i = 0; i < n; ++i)
     {
